refactor(ios): move doctor-summary aggregators to an extension file
The new helpers pushed AnalyticsInsights.swift past the 1000-line
file_length limit and the enum body past the 800-line type_body_length
limit (both error-severity in CI). Move monthlyHeadacheDays,
medicationUsage, and preventativeCompliance into
AnalyticsInsights+DoctorSummary.swift. No behavior change.
